+++ packages/tenbo/lib/help.mjs
@@ -8,6 +8,7 @@ Options:
   -l|--launcher <module>  puppeteer module name (default: puppeteer)
   -e|--executable <path>  override path to the Chromium executable
   -t|--timeout <ms>       fail if no log or request occurs (default: 1000)
+  -c|--cache <s>          enable browser cache (default: 5)
   -a|--[no-]headless      enable headless mode (default: true)
   -d|--[no-]disconnect    just disconnect from the browser at the end
   -v|--[no-]verbose       print more diagnostics on the console